moving to capitalized aliases convention showed by webpack docs, changing components and containers to index.jsx, reworked the aliases a bit and added resolve paths which WP2 requires, updated extensions so index can be used, updated css-loader and url-resolve-loader versions,
