+++ app/components/header/__tests__/header.spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { mount } from 'enzyme';
 import { expect } from 'chai';
 
-import Header from '../header';
+import Header from '../index';
 
 describe('Header component', () => {
