+++ app/components/warrior/__tests__/warrior.spec.jsx
@@ -2,13 +2,13 @@ import React from 'react';
 import { mount } from 'enzyme';
 import chai from 'chai';
 import chaiEnzyme from 'chai-enzyme';
-import { IMAGE_PATH } from 'constants/appConstants';
+import { IMAGE_PATH } from 'Constants/appConstants';
 
 chai.use(chaiEnzyme());
 
 const expect = chai.expect;
 
-import Warrior from '../warrior';
+import Warrior from '../index';
 
 describe('The Warrior component', () => {
