+++ app/components/warrior/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import {IMAGE_PATH} from 'constants/appConstants';
+import {IMAGE_PATH} from 'Constants/appConstants';
 
 import 'dmc/css/components/warrior.min.css';
