+++ app/containers/warriorsList/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {connect} from 'react-redux';
-import Warrior from 'warrior';
-import {IMAGE_PATH} from 'constants/appConstants';
+import Warrior from 'Components/warrior';
+import {IMAGE_PATH} from 'Constants/appConstants';
 import 'dmc/css/components/warriorsList.min.css';
 import PropTypes from 'prop-types';
