+++ app/main.jsx
@@ -5,9 +5,9 @@ import { createLogger } from 'redux-logger';
 import { createStore, applyMiddleware } from 'redux';
 import reducer  from './reducers/index.js';
 import io from 'socket.io-client';
-import { receiveWarriors, chooseOpponents, notify } from 'actions';
+import { receiveWarriors, chooseOpponents, notify } from 'Actions';
 import { Provider } from 'react-redux';
-import Content from 'content';
+import Content from 'Components/content';
 
 // ================================================================
 // use the logger middleware
