+++ package.json
@@ -29,7 +29,7 @@
     "babel-preset-react": "^6.5.0",
     "chai": "^3.5.0",
     "chai-enzyme": "^0.5.0",
-    "css-loader": "^0.25.0",
+    "css-loader": "^0.28.4",
     "enzyme": "^2.3.0",
     "express": "^4.13.4",
     "gulp": "^3.9.1",
@@ -54,13 +54,13 @@
     "phantomjs-polyfill-object-assign": "0.0.2",
     "react-addons-test-utils": "^15.1.0",
     "redux-devtools-extension": "^2.13.2",
-    "resolve-url-loader": "^1.6.0",
+    "resolve-url-loader": "^2.1.0",
     "sass-loader": "^4.0.2",
     "sinon": "^1.17.4",
     "socket.io": "^1.4.6",
     "style-loader": "^0.13.1",
     "svg-url-loader": "^1.1.0",
-    "webpack": "^1.13.1"
+    "webpack": "^2.6.1"
   },
   "license": "ISC"
 }
