+++ webpack/extensions.js
@@ -1,1 +1,1 @@
-module.exports = ['', '.js', '.jsx', '.svg'];
+module.exports = ['*', '.js', '.jsx', '.svg'];
